backend/erp/requirements.txt
PyJWT==2.8.0
Django==5.2.11
mysqlclient
